Created generate script
